Optimize MjpegDataProvider: One http connection for one show session and break stream when data show pause

+++ MjpegStreamReciever/MjpegStreamReciever/Model/DataProviders/MjpegDataProvider.cs
@@ -1,11 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Drawing;
 using System.IO;
 using System.Linq;
 using System.Net;
 using System.Net.Http;
 using System.Text;
+using System.Threading;
 using System.Threading.Tasks;
 using System.Windows.Media.Imaging;
 
@@ -13,18 +15,57 @@ namespace MjpegStreamReciever.Model.DataProviders
 {
     public class MjpegDataProvider : DataProvider
     {
-        private string lastSource;
+        private object locker = new object();
+        private string lastSource = "";
         private byte[] lastDataFrame;
+        private Stopwatch timer = new Stopwatch();
+        private CancellationTokenSource cts = new CancellationTokenSource();
+        private const int timeout = 5000;
+       
         public override byte[] Provide(string source)
         {
+
+            if (!lastSource.Equals(source) || isInited)
+            {
+                isInited = false;
+                lastSource = source;
+                cts.Cancel();
+                cts = new CancellationTokenSource();
+                StartParseStream(source);
+
+            }
+            lock (locker)
+            {
+                timer.Restart();
+            }
+            return lastDataFrame;
+        }
+        private void StartParseStream(string source)
+        {
+            HttpClient www = new HttpClient();
+            var progress = new Progress<byte[]>(x => lastDataFrame = x);
+            Task.Run(() => ParseStreamAsync(progress, www.GetStreamAsync(source), cts.Token, timeout));
+        }
+        private async Task ParseStreamAsync(IProgress<byte[]> progress,
+            Task<Stream> streamProvider, CancellationToken token, int timeout)
+        {
+            long elapsed = 0;
+            lock (locker)
+            {
+                elapsed = timer.ElapsedMilliseconds;
+            }
+            if (elapsed > timeout)
+            {
+                cts.Cancel();
+            }
             List<int> last4Bytes = new List<int> { 0, 0, 0, 0 };
             List<byte> data = new List<byte>();
-            HttpClient www = new HttpClient();
-            string esc = "\r\n\r\n";
-            bool flag = false;
             bool started = false;
 
-            using (var stream = www.GetStreamAsync(source).Result)
+            string esc = "\r\n\r\n";
+            await streamProvider;
+
+            using (var stream = streamProvider.Result)
             {
                 while (stream.CanRead)
                 {
@@ -32,16 +73,30 @@ namespace MjpegStreamReciever.Model.DataProviders
                     if (currentByte == -1) break;
                     last4Bytes.RemoveAt(0);
                     last4Bytes.Add(currentByte);
-                    flag = IsEscape(last4Bytes, esc);
+                    bool flag = IsEscape(last4Bytes, esc);
 
                     if (flag)
                     {
                         if (started)
                         {
+                            lock (locker)
+                            {
+                                elapsed = timer.ElapsedMilliseconds;
+                            }
+                            if (elapsed > timeout || token.IsCancellationRequested)
+                            {
+
+                                progress.Report(null);
+                                return;
+                            }
                             int needToRemove = esc.Length - 1;
                             data.RemoveRange(data.Count - needToRemove - 1, needToRemove);
-                            var str = Encoding.Default.GetString(data.ToArray());
-                            return data.ToArray();
+
+                            progress.Report(data.ToArray());
+                            last4Bytes = new List<int> { 0, 0, 0, 0 };
+                            data = new List<byte>();
+                            started = false;
+
                         }
                         else
                         {
@@ -55,9 +110,9 @@ namespace MjpegStreamReciever.Model.DataProviders
                         data.Add((byte)currentByte);
                     }
                 }
-
             }
-            return null;
+
+
         }
         private bool IsEscape(List<int> bytes, string esc)
         {
@@ -70,6 +125,5 @@ namespace MjpegStreamReciever.Model.DataProviders
             }
             return true;
         }
-
     }
 }
